Hvbassist/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
firebase-admin
